Test auto discovery
