+++ src/image_builder/tests/02_integration/build_params/test_01_tar.json
