+++ src/image_builder/tests/02_integration/build_params/test_02_no_context.name
