+++ src/image_builder/tests/02_integration/build_params/test_03_build_context.json
