+++ src/image_builder/tests/02_integration/build_params/test_04_image_variants.json
@@ -1,0 +1,17 @@
+{
+  "source_type": "git",
+  "source_repo": {
+      "url": "https://github.com/mihaTrajbaric/generic_repo.git"
+    },
+  "target_images": [
+    {
+      "image": "tests/image_variants",
+      "tag": "latest"
+    },
+    {
+      "image": "tests/image_variants",
+      "tag": "python-3.8",
+      "base": "python:3.8-alpine"
+    }
+  ]
+}
