+++ src/image_builder/tests/02_integration/build_params/test_04_image_variants.name
