+++ src/image_builder/tests/02_integration/build_params/test_05_subdir_context.json
