+++ src/image_builder/tests/02_integration/build_params/test_06_git.json
