+++ src/image_builder/tests/02_integration/build_params/test_07_git_workdir_dockerfile.json
