--- src/image_builder/tests/integration/build_params/test_04_image_variants.json
@@ -1,18 +1,0 @@
-{
-  "source_type": "dockerfile",
-  "source_url": "https://raw.githubusercontent.com/mihaTrajbaric/image-builder-test-files/master/image-variants/Dockerfile",
-  "build_context": {
-      "url": "https://github.com/mihaTrajbaric/generic_docker_build_context.git"
-    },
-  "target_images": [
-    {
-      "image": "tests/image_variants",
-      "tag": "latest"
-    },
-    {
-      "image": "tests/image_variants",
-      "tag": "python-3.8",
-      "base": "python:3.8-alpine"
-    }
-  ]
-}
